Addressing the improvements within the ApplicationDecorator and restructuring the view templates to use a partial for imported attributes
