+++ app/views/catalog/_resource_attributes_default.html.erb
@@ -1,3 +1,6 @@
+<% if document.decorated_resource.try(:imported_attributes) %>
+  <%= render 'resource_imported_attributes', document: document %>
+<% end %>
 <div id="attributes">
   <table class="table table-striped <%= document.decorated_resource.model_name.singular %> attributes">
     <caption class="table-heading">
