--- app/views/catalog/_resource_attributes_ephemera_folder.html.erb
@@ -1,28 +1,0 @@
-<div id="attributes">
-  <table class="table table-striped <%= document.decorated_resource.model_name.singular %> attributes">
-    <caption class="table-heading">
-      <h2>Attributes</h1>
-    </caption>
-    <thead>
-      <tr>
-        <th>Attribute Name</th>
-        <th>Values</th>
-      </tr>
-    </thead>
-    <tbody>
-      <% document.decorated_resource.display_attributes.each do |attribute, values| %>
-        <% if values.present? %>
-          <tr>
-            <th><%= t("valhalla.works.show.attributes.#{attribute}.label", default: attribute.to_s.titleize) %></th>
-            <td>
-              <ul class="tabular">
-                <% values.each do |value| %>
-                  <li class="attribute <%= attribute %>" dir="<%= value.to_s.dir %>"><%= value %></li>
-                <% end %>
-              </ul>
-          </tr>
-        <% end %>
-      <% end %>
-    </tbody>
-  </table>
-</div>
