+++ spec/support/capybara_selenium.rb
@@ -4,7 +4,7 @@ require 'selenium-webdriver'
 
 Capybara.register_driver(:headless_chrome) do |app|
   capabilities = Selenium::WebDriver::Remote::Capabilities.chrome(
-    chromeOptions: { args: %w(headless disable-gpu disable-setuid-sandbox) }
+    chromeOptions: { args: %w(headless disable-gpu disable-setuid-sandbox window-size=7680,4320) }
   )
 
   Capybara::Selenium::Driver.new(app,
